move auth client to the features/auth
