+++ src/features/auth/forms/signup-form.tsx
@@ -15,7 +15,7 @@ import {
   FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-import { authClient } from "@/lib/auth-client";
+import { authClient } from "@/features/auth/auth-client";
 import { useState } from "react";
 
 const formSchema = v.object({
