+++ src/features/auth/index.ts
@@ -5,13 +5,7 @@ export { ForgotPasswordForm } from "./forms/forgot-password-form";
 export { ResetPasswordForm } from "./forms/reset-password-form";
 
 // Components
-export { LogoutButton } from "./logout-button";
+export { LogoutButton } from "./components/logout-button";
 
 // Hooks (when we create them)
 // export { useAuth } from "./hooks/use-auth";
-
-// Lib (when we move auth-client)
-// export { authClient } from "./lib/auth-client";
-
-// Providers (when we move convex-provider)
-// export { ConvexClientProvider } from "./providers/convex-provider";
